Change redirect URL when that beatmap is already approve

+++ beatmap_collections/views.py
@@ -131,7 +131,7 @@ def approve_beatmap(request, collection_id, beatmap_entry_id):
         return redirect('collection', collection_id=collection_id)
     if beatmap_entry.owner_approved:
         messages.error(request, 'This beatmap is already approved!')
-        return redirect('collection', collection_id=collection_id)
+        return redirect('beatmap_approval', collection_id=collection_id)
     beatmap_entry.owner_approved = True
     beatmap_entry.save()
     messages.success(request, 'Beatmap approved!')
@@ -146,6 +146,9 @@ def deny_beatmap(request, collection_id, beatmap_entry_id):
     if request.user != collection.author:
         messages.error(request, 'Hehehehe No! Stop there!')
         return redirect('collection', collection_id=collection_id)
+    if beatmap_entry.owner_approved:
+        messages.error(request, 'This beatmap is already approved!')
+        return redirect('beatmap_approval', collection_id=collection_id)
     beatmap_entry.delete()
     messages.success(request, 'Beatmap denied!')
     return redirect('beatmap_approval', collection_id=collection_id)
